fix signed/unsigned mismatch C4018 on win x86 auditors @bbondy

+++ atom/common/atom_command_line.cc
@@ -19,7 +19,7 @@ AtomCommandLine::StringVector AtomCommandLine::argv_utf8_;
 // static
 void AtomCommandLine::InitializeFromCommandLine() {
   CommandLine::StringVector argv = CommandLine::ForCurrentProcess()->argv();
-  for (int i = 0; i < argv.size(); ++i) {
+  for (size_t i = 0; i < argv.size(); ++i) {
 #if defined(OS_WIN)
     argv_utf8_.push_back(base::SysWideToUTF8(argv[i]));
 #else
